backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
google-genai==0.3.0
python-dotenv==1.0.1
pandas==2.2.3
